+++ src/components/calcite-input-message/calcite-input-message.e2e.ts
@@ -1,0 +1,87 @@
+import { newE2EPage } from "@stencil/core/testing";
+
+describe("calcite-input-message", () => {
+  it("renders", async () => {
+    const page = await newE2EPage();
+    await page.setContent("<calcite-input-message></calcite-input-message>");
+    const input = await page.find("calcite-input-message");
+    expect(input).toHaveClass("hydrated");
+  });
+
+  it("renders default props when none are provided", async () => {
+    const page = await newE2EPage();
+    await page.setContent(`
+    <calcite-input-message></calcite-input-message>
+    `);
+
+    const element = await page.find("calcite-input-message");
+    expect(element).toEqualAttribute("status", "idle");
+    expect(element).toEqualAttribute("theme", "light");
+    expect(element).toEqualAttribute("appearance", "default");
+    expect(element).toEqualAttribute("type", "default");
+  });
+
+  it("renders default props when invalid props are provided", async () => {
+    const page = await newE2EPage();
+    await page.setContent(`
+    <calcite-input-message status="zip" theme="zap" appearance="zom" type="zur"></calcite-input-message>
+    `);
+
+    const element = await page.find("calcite-input-message");
+    expect(element).toEqualAttribute("status", "idle");
+    expect(element).toEqualAttribute("theme", "light");
+    expect(element).toEqualAttribute("appearance", "default");
+    expect(element).toEqualAttribute("type", "default");
+  });
+
+  it("renders requested props when valid props are provided", async () => {
+    const page = await newE2EPage();
+    await page.setContent(`
+    <calcite-input-message status="valid" theme="dark" appearance="minimal" type="floating"></calcite-input-message>
+    `);
+
+    const element = await page.find("calcite-input-message");
+    expect(element).toEqualAttribute("status", "valid");
+    expect(element).toEqualAttribute("theme", "dark");
+    expect(element).toEqualAttribute("appearance", "minimal");
+    expect(element).toEqualAttribute("type", "floating");
+  });
+
+  it("inherits requested props when from wrapping calcite-label when props are provided", async () => {
+    const page = await newE2EPage();
+    await page.setContent(`
+    <calcite-label status="invalid" theme="dark">
+    Label text
+    <calcite-input-message></calcite-input-message>
+    </calcite-label>
+    `);
+
+    const element = await page.find("calcite-input-message");
+    expect(element).toEqualAttribute("status", "invalid");
+    expect(element).toEqualAttribute("theme", "dark");
+  });
+
+  it("does not render an icon if not requested", async () => {
+    const page = await newE2EPage();
+    await page.setContent(`
+    <calcite-input-message></calcite-input-message>
+    `);
+
+    const icon = await page.find(
+      "calcite-input-message >>> .calcite-input-message-icon"
+    );
+    expect(icon).toBeNull();
+  });
+
+  it("renders an icon if requested", async () => {
+    const page = await newE2EPage();
+    await page.setContent(`
+    <calcite-input-message icon></calcite-input-message>
+    `);
+
+    const icon = await page.find(
+      "calcite-input-message >>> .calcite-input-message-icon"
+    );
+    expect(icon).not.toBeNull();
+  });
+});