+++ CHANGELOG.md
@@ -8,17 +8,20 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## Unreleased
 
 ### Breaking Changes
+
 - `calcite-button` no longer accepts `inline` as a value for `appearance` - you can instead
 
 ### Added
 
 - new component `calcite-link`
+- new `calcite-label`, `calcite-input`, and `calcite-input-message` components
+- `calcite-slider` can now be programmatically focused with the `setFocus()` method
 
 ## [v1.0.0-beta.24] - Apr 8th 2020
 
 ### Fixed
 
-- fix NPM deployment issue
+- fix NPM release issue
 
 ## [v1.0.0-beta.23] - Apr 7th 2020